check_folds {stabs}	R Documentation

Check if folds result from subsampling with p = 0.5

Description

(Internal) function that checks if folds result from subsampling with p = 0.5 and adds complementary pairs if needed.

Usage

check_folds(folds, B, n, sampling.type)

Arguments

folds
a weight matrix that represents the subsamples.

B
number of subsampling replicates.

n
the number of observations; needed for internal checks.

sampling.type
sampling type to be used.

Details

This is an internal function used to check if folds are specified correctly. For details (e.g. on arguments) see stabsel.

Value

A matrix containing the folds, possibly after adding the complementary pairs.

References

B. Hofner, L. Boccuto and M. Goeker (2015), Controlling false discoveries in high-dimensional situations: Boosting with stability selection. BMC Bioinformatics, 16:144.
doi: 10.1186/s12859-015-0575-3.

See Also

For details see stabsel.

[Package stabs version 0.7-1 Index]