stabsel.stabsel {stabs}	R Documentation

Change Parameters of Stability Selection

Description

Method to change the parameters cutoff, PFER and assumption of stability selection that can be altered without the need to re-run the subsampling process.

Usage

## S3 method for class 'stabsel'
stabsel(x, cutoff, PFER, assumption = x$assumption, ...)

Arguments

x
an object that results from a call to stabsel.

cutoff
cutoff between 0.5 and 1. Preferably a value between 0.6 and 0.9 should be used. If assumption = "r-concave" the cutoff can be below 0.5 but needs to be larger than q/p.

PFER
upper bound for the per-family error rate. This specifies the amount of falsely selected base-learners, which is tolerated. See details.

assumption
Defines the type of assumptions on the distributions of the selection probabilities and simultaneous selection probabilities. Only applicable for sampling.type = "SS". For sampling.type = "MB" we always use "none".

...
additional arguments that are currently ignored.

Details

This function allows to alter the parameters cutoff, PFER and assumption of a fitted stability selection result. All other parameters are re-used from the original stability selection results. The missing paramter is computed and the selected variables are updated accordingly.

Value

An object of class stabsel. For details see there.

See Also

stabsel for the generic function, stabsel_parameters for the computation of error bounds, fitfun for available fitting functions and plot.stabsel for available plot functions

Examples

  if (require("TH.data")) {
      ## make data set available
      data("bodyfat", package = "TH.data")
  } else {
      ## simulate some data if TH.data not available.
      ## Note that results are nonsense with this data.
      bodyfat <- matrix(rnorm(720), nrow = 72, ncol = 10)
  }

  ## set seed
  set.seed(1234)

  ####################################################################
  ### using stability selection with Lasso methods:

  if (require("lars")) {
      (stab.lasso <- stabsel(x = bodyfat[, -2], y = bodyfat[,2],
                             fitfun = lars.lasso, cutoff = 0.75,
                             PFER = 1, mc.cores = 1L))
      ## set mc.cores = 1L only to avoid parallelization on CRAN;
      ## In practice you usually do not need set this value unless you want to
      ## avoid parallelization via mclapply or restrict it to a specific number
      ## of available cores. See documentation of mclapply for more details.

      par(mfrow = c(2, 1))
      plot(stab.lasso)

      ## now change the PFER and the assumption:
      (stab.lasso_cf0.93_rconc <- stabsel(stab.lasso, cutoff = 0.93,
                                          assumption = "r-concave", mc.cores = 1L))
      ## In practice you usually do not need set this value unless you want to
      ## avoid parallelization via mclapply or restrict it to a specific number
      ## of available cores. See documentation of mclapply for more details.

      plot(stab.lasso_cf0.93_rconc)
      ## the cutoff did change and hence the PFER and the selected
      ## variables
  }

[Package stabs version 0.7-1 Index]