subsample {stabs}	R Documentation

Draw Random Subsamples

Description

Set up weight matrix for subsampling with sample proportion 1/2 to be used with stabsel.

Usage

subsample(weights, B = 100, strata = NULL)

Arguments

weights
a numeric vector of weights for the model to be cross-validated.

B
number of folds, per default 25 for bootstrap and subsampling and 10 for kfold.

strata
a factor of the same length as weights for stratification.

Details

The function subsample can be used to build an appropriate weight matrix to be used with stabsel. See there for more details.

If strata is defined sampling is performed in each stratum separately thus preserving the distribution of the strata variable in each fold.

See Also

stabsel

Examples

  ## just a low-dimensional example
  subsample(weights = rep(1, 10), B = 50)

[Package stabs version 0.7-1 Index]