run_stabsel {stabs}	R Documentation

Run Stability Selection

Description

(Internal) function that is used to run stability selection (i.e. to apply the fit-function to the subsamples. This function is not intended to be directly called.

Usage

run_stabsel(fitter, args.fitter, n, p, cutoff, q, PFER, folds, B, assumption,
            sampling.type, papply, verbose, FWER, eval, names, ...)

Arguments

fitter
a function to fit the model on subsamples. See argument fitfun of stabsel for details.

args.fitter
a named list containing additional arguments that are passed to fitter. See argument args.fitfun stabsel for details.

n
the number of observations; needed for internal checks.

p
number of possible predictors (including intercept if applicable).

cutoff
cutoff between 0.5 and 1.

q
number of (unique) selected variables (or groups of variables depending on the model) that are selected on each subsample.

PFER
upper bound for the per-family error rate.

folds
a weight matrix that represents the subsamples.

B
number of subsampling replicates.

assumption
distributional assumption.

sampling.type
sampling type to be used.

papply
(parallel) apply function.

verbose
logical (default: TRUE) that determines wether warnings should be issued.

FWER
deprecated. Only for compatibility with older versions, use PFER instead.

eval
logical. Determines whether stability selection is evaluated.

names
variable names that are used to label the results.

...
additional arguments to be passed to next function.

Details

This is an internal function that fits the actual models to the subsamples, i.e., this is the work horse that runs stability selection. Usually, one should use stabsel, which internally calls run_stabsel.

run_stabsel can be used by expert users to implement stability selection methods for new model types.

For details (e.g. on arguments) see stabsel.

Value

An object of class stabsel with the following elements:

phat
selection probabilities.

selected
elements with maximal selection probability greater cutoff.

max
maximum of selection probabilities.

cutoff
cutoff used.

q
average number of selected variables used.

PFER
per-family error rate.

p
the number of effects subject to selection.

sampling.type
the sampling type used for stability selection.

assumption
the assumptions made on the selection probabilities.

References

B. Hofner, L. Boccuto and M. Goeker (2014), Controlling false discoveries in high-dimensional situations: Boosting with stability selection. Technical Report, arXiv:1411.1285.
http://arxiv.org/abs/1411.1285.

See Also

For details see stabsel.

[Package stabs version 0.5-1 Index]