stabsel_parameters {stabs}	R Documentation

Compute Error Bounds for Stability Selection

Description

Compute the missing parameter from the two given parameters in order to assess suitability of the parameter constellation

Usage

stabsel_parameters(p, ...)

## Default S3 method:
stabsel_parameters(p, cutoff, q, PFER,
                   B = ifelse(sampling.type == "MB", 100, 50),
                   assumption = c("unimodal", "r-concave", "none"),
                   sampling.type = c("SS", "MB"),
                   verbose = FALSE, FWER, ...)

## S3 method for class 'stabsel_parameters'
print(x, heading = TRUE, ...)

Arguments

p
number of possible predictors (including intercept if applicable).

cutoff
cutoff between 0.5 and 1. Preferably a value between 0.6 and 0.9 should be used.

q
number of (unique) selected variables (or groups of variables depending on the model) that are selected on each subsample.

PFER
upper bound for the per-family error rate. This specifies the amount of falsely selected base-learners, which is tolerated. See details.

B
number of subsampling replicates. Per default, we use 50 complementary pairs for the error bounds of Shah & Samworth (2013) and 100 for the error bound derived in Meinshausen & Buehlmann (2010). As we use B complementray pairs in the former case this leads to 2B subsamples.

assumption
Defines the type of assumptions on the distributions of the selection probabilities and simultaneous selection probabilities. Only applicable for sampling.type = "SS". For sampling.type = "MB" we always use code"none".

sampling.type
use sampling scheme of of Shah & Samworth (2013), i.e., with complementarty pairs (sampling.type = "SS"), or the original sampling scheme of Meinshausen & Buehlmann (2010).

verbose
logical (default: TRUE) that determines wether warnings should be issued.

FWER
deprecated. Only for compatibility with older versions, use PFER instead.

x
an object of class "stabsel_parameters".

heading
logical. Specifies if a heading line should be printed.

...
additional arguments to be passed to next function.

Details

This function implements the error bounds for stability selection by Meinshausen and Buehlmann (2010) and the improved error bounds by Shah and Samworth (2013). For details see also Hofner et al. (2014).

Two of the three arguments cutoff, q and PFER must be specified. The per-family error rate (PFER), i.e., the expected number of false positives E(V), where V is the number of false positives, is bounded by the argument PFER.

For more details see also stabsel.

Value

An object of class stabsel_parameters with a special print method. The object has the following elements:

cutoff
cutoff used.

q
average number of selected variables used.

PFER
(realized) upper bound for the per-family error rate.

specifiedPFER
specified upper bound for the per-family error rate.

p
the number of effects subject to selection.

B
the number of subsamples.

sampling.type
the sampling type used for stability selection.

assumption
the assumptions made on the selection probabilities.

References

B. Hofner, L. Boccuto and M. Goeker (2014), Controlling false discoveries in high-dimensional situations: Boosting with stability selection. Technical Report, arXiv:1411.1285.
http://arxiv.org/abs/1411.1285.

N. Meinshausen and P. Buehlmann (2010), Stability selection. Journal of the Royal Statistical Society, Series B, 72, 417–473.

R.D. Shah and R.J. Samworth (2013), Variable selection with error control: another look at stability selection. Journal of the Royal Statistical Society, Series B, 75, 55–80.

See Also

For more details see also stabsel.

[Package stabs version 0.5-1 Index]